+++ src/main/java/com/sparta/todayeats/address/entity/Address.java
@@ -1,6 +1,6 @@
 package com.sparta.todayeats.address.entity;
 
-import com.sparta.todayeats.address.dto.reqeust.AddressUpdateRequest;
+import com.sparta.todayeats.address.dto.request.AddressUpdateRequest;
 import com.sparta.todayeats.global.infrastructure.entity.BaseEntity;
 import com.sparta.todayeats.user.entity.User;
 import jakarta.persistence.*;
